requirements.txt
Flask==2.3.2
pandas==2.2.2
numpy==1.26.4
pillow==10.0.1
pytesseract==0.3.12
nltk==3.9.0
scikit-learn==1.3.2
werkzeug==3.0.1
